pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "business-match-bot"
version = "1.0.0"
description = "Telegram bot for business networking and professional matching"
authors = [{name = "Business Match"}]
requires-python = ">=3.8"
dependencies = [
    "python-telegram-bot==20.8",
    "pandas==2.1.4",
    "openai==1.3.7",
    "python-dotenv==1.0.0",
    "openpyxl==3.1.2",
    "flask==3.0.0",
    "requests==2.31.0",
]

[project.scripts]
bot = "main:main"

